requirements.txt
numpy
opencv-python
pillow
tensorflow
pytesseract
easyocr
jiwer
pandas
seaborn
matplotlib